Reorganized job and execution configuration files
