--- scale/job/configuration/execution/exceptions.py
@@ -1,8 +1,0 @@
-"""Defines exceptions that can occur when interacting with execution configuration"""
-
-
-class InvalidExecutionConfiguration(Exception):
-    """Exception indicating that the provided execution configuration JSON was invalid
-    """
-
-    pass
