+++ scale/job/configuration/interface/job_interface.py
@@ -13,7 +13,7 @@ from job.configuration.data.exceptions import InvalidData, InvalidConnection
 from job.configuration.interface import job_interface_1_3 as previous_interface
 from job.configuration.interface.exceptions import InvalidInterfaceDefinition
 from job.configuration.interface.scale_file import ScaleFileDescription
-from job.configuration.job.exceptions import MissingSetting
+from job.configuration.exceptions import MissingSetting
 from job.configuration.results.exceptions import InvalidResultsManifest
 from job.configuration.results.results_manifest.results_manifest import ResultsManifest
 from job.execution.container import SCALE_JOB_EXE_INPUT_PATH, SCALE_JOB_EXE_OUTPUT_PATH
@@ -497,7 +497,7 @@ class JobInterface(object):
         :param command_arguments: The command_arguments that you want to perform the replacement on
         :type command_arguments: string
         :param job_configuration: The job configuration
-        :type job_configuration: :class:`job.configuration.execution.json.exe_config.ExecutionConfiguration`
+        :type job_configuration: :class:`job.configuration.json.execution.exe_config.ExecutionConfiguration`
         :param job_type: The job type definition 
         :type job_type: :class:`job.models.JobType`
         :return: command arguments with the settings populated
@@ -518,7 +518,7 @@ class JobInterface(object):
         """Populates the environment variables with the requested values.
 
         :param job_configuration: The job configuration
-        :type job_configuration: :class:`job.configuration.execution.json.exe_config.ExecutionConfiguration`
+        :type job_configuration: :class:`job.configuration.json.execution.exe_config.ExecutionConfiguration`
         :param job_type: The job type definition 
         :type job_type: :class:`job.models.JobType`
 
@@ -577,7 +577,7 @@ class JobInterface(object):
         """Ensures that all required settings are defined in the job_configuration
 
         :param job_configuration: The job configuration
-        :type job_configuration: :class:`job.configuration.execution.json.exe_config.ExecutionConfiguration`
+        :type job_configuration: :class:`job.configuration.json.execution.exe_config.ExecutionConfiguration`
         """
 
         interface_settings = self.definition['settings']
@@ -740,7 +740,7 @@ class JobInterface(object):
         :param settings: The job configuration
         :type settings: JSON
         :param job_configuration: The job configuration
-        :type job_configuration: :class:`job.configuration.execution.json.exe_config.ExecutionConfiguration`
+        :type job_configuration: :class:`job.configuration.json.execution.exe_config.ExecutionConfiguration`
         :return: settings name and the value to replace it with
         :rtype: dict
         """
