+++ scale/job/test/configuration/interface/test_job_interface.py
@@ -11,10 +11,10 @@ import storage.test.utils as storage_test_utils
 from job.configuration.data.exceptions import InvalidConnection, InvalidData
 from job.configuration.data.job_connection import JobConnection
 from job.configuration.data.job_data import JobData
-from job.configuration.execution.json.exe_config import ExecutionConfiguration
 from job.configuration.interface.exceptions import InvalidInterfaceDefinition
 from job.configuration.interface.job_interface import JobInterface
-from job.configuration.job.exceptions import MissingSetting
+from job.configuration.exceptions import MissingSetting
+from job.configuration.json.execution.exe_config import ExecutionConfiguration
 from job.configuration.results.exceptions import InvalidResultsManifest
 from job.execution.container import SCALE_JOB_EXE_INPUT_PATH, SCALE_JOB_EXE_OUTPUT_PATH
