+++ scale/queue/models.py
@@ -11,7 +11,7 @@ from django.db import models, transaction
 from error.models import Error
 from job.configuration.data.exceptions import InvalidData
 from job.configuration.data.job_data import JobData
-from job.configuration.execution.json.exe_config import ExecutionConfiguration
+from job.configuration.json.execution.exe_config import ExecutionConfiguration
 from job.execution.job_exe import RunningJobExecution
 from job.models import Job, JobType
 from job.models import JobExecution
@@ -440,7 +440,7 @@ class QueueManager(models.Manager):
         :param event: The event that triggered the creation of this job
         :type event: :class:`trigger.models.TriggerEvent`
         :param configuration: The optional initial execution configuration
-        :type configuration: :class:`job.configuration.execution.json.exe_config.ExecutionConfiguration
+        :type configuration: :class:`job.configuration.json.execution.exe_config.ExecutionConfiguration`
         :returns: The new queued job
         :rtype: :class:`job.models.Job`
